김세형 주문, 리뷰, 환불, 내정보, 대시보드, 주문관리, 리뷰관리, 환불관리, 회원관리

+++ src/main/webapp/mypage_review/review_delete.jsp
@@ -1,6 +1,11 @@
 <%@ page contentType="text/html; charset=UTF-8" pageEncoding="UTF-8" %>
 <%@ page import="review.*" %>
 <%
+
+if (session.getAttribute("userId") ==null ) {
+	  response.sendRedirect("/mall_prj/UserLogin/login.jsp");
+	  return;
+	}
   int reviewId = Integer.parseInt(request.getParameter("review_id"));
   int result = new ReviewService().deleteReview(reviewId);
 %>
